아... 나머지연산.. 왜까먹었지..

+++ app/src/main/java/com/example/mobile_programming_term_project/makeCalculationStack.java
@@ -4,6 +4,6 @@ public interface makeCalculationStack {
     boolean isEmpty();
     boolean isFull();
     void push(int item);
-    int pop();
+    double pop();
     int peek();
 }
